Windows: Start refactor execdriver/driver.go
